// src/utils/types/GeoJSON.ts
export declare namespace GeoJSON {
// ...
	export type Position = [longitude: number, latitude: number, elevation?: number];

	export type Record = { [key in string | number]: unknown };

	/**
	 * Properties inherit to all GeoJSON types
	 */
	export interface GeometryBase extends Record {
		/**
		 * A GeoJSON object MAY have a member named "bbox" to include information on the coordinate range for its
		 * Geometries, Features, or FeatureCollections. The value of the bbox member MUST be an array of length 2*n
		 * where n is the number of dimensions represented in the contained geometries, with all axes of the most
		 * southwesterly point followed by all axes of the more northeasterly point. The axes order of a bbox follows
		 * the axes order of geometries.
		 */
		bbox?: number[];

		/**
		 * A GeoJSON object MAY have other members.
		 *
		 * Members not described in this specification ("foreign members") MAY be used in a GeoJSON document. Note that
		 * support for foreign members can vary across implementations, and no normative processing model for foreign
		 * members is defined.
		 */
	}

	// geometry types

	export interface Point extends GeometryBase {
		type: "Point";
		/**
		 * For type "Point", the "coordinates" member is a single position.
		 */
		coordinates: Position;
	}
// ...
}